+++ djangocms_content_expiry/test_utils/polls/admin.py
@@ -1,5 +1,5 @@
-from django.conf.urls import url
 from django.contrib import admin
+from django.urls import re_path
 
 from .models import Answer, Poll, PollContent
 from .views import PreviewView
@@ -10,7 +10,7 @@ class PollContentAdmin(admin.ModelAdmin):
     def get_urls(self):
         info = self.model._meta.app_label, self.model._meta.model_name
         return [
-            url(
+            re_path(
                 r"^(?P<id>\d+)/preview/$",
                 self.admin_site.admin_view(PreviewView.as_view()),
                 name="{}_{}_preview".format(*info),