+++ djangocms_content_expiry/test_utils/polls/models.py
@@ -20,10 +20,10 @@ class PollContent(models.Model):
         return self.text
 
     def get_absolute_url(self):
-        return reverse("admin:polls_pollcontent_change", args=[self.id])
+        return reverse(f"admin:{self._meta.app_label}_{self._meta.model_name}_change", args=[self.id])
 
     def get_preview_url(self):
-        return reverse("admin:polls_pollcontent_preview", args=[self.id])
+        return reverse(f"admin:{self._meta.app_label}_{self._meta.model_name}_preview", args=[self.id])
 
 
 class Answer(models.Model):